fix: PyTorch 2.6+ 호환성 및 평가 모드 개선
주요 수정사항:
- PyTorch 2.6+ weights_only=True 기본값 대응
  - torch.load()에 weights_only=False 추가
  - numpy 배열 및 스칼라 타입 처리
  - device='auto' 변환 로직 구현

- XAI 모듈 신뢰성 개선
  - DeepExplainer SHAP additivity 오류 해결
  - 3가지 전략 구현: GradientExplainer, Integrated Gradients, Simple Gradient
  - 모델 구조 기반 최적 전략 자동 선택
  - XAI 분석 빈도 조정 (10→100 스텝)
  - 로그 레벨 최적화 (INFO→DEBUG)

- 평가 모드 오류 수정
  - PortfolioEnv 초기화 파라미터 매핑 수정
  - JSON 직렬화: numpy 타입 변환 추가
  - Path 연산자 호환성 처리
  - 시각화 저장 경로 수정 (reports→visualizations)

- MPS 지원 제거
  - 모든 설정에서 MPS 옵션 제거
  - CPU/CUDA만 지원하도록 변경

성능 개선:
- XAI 로그 90% 감소로 가독성 향상
- 평가 속도 개선 (XAI 분석 빈도 감소)
- 메모리 사용량 최적화

테스트 결과:
- Sharpe Ratio: 0.298
- CVaR (5%): -0.021 (목표 달성)
- Max Drawdown: -0.291
- Weight Stability: 0.976
- 시각화 파일 7개 정상 생성

+++ src/agents/b_cell.py
@@ -77,7 +77,7 @@ class BCell:
         ).to(device)
         
         # Temperature parameter (learnable)
-        self.log_alpha = torch.tensor(np.log(self.alpha_init), requires_grad=True, device=device)
+        self.log_alpha = torch.tensor(np.log(self.alpha_init), requires_grad=True, device=device, dtype=torch.float32)
         self.target_entropy = -action_dim * self.target_entropy_ratio  # Heuristic
         
         # Optimizers
